Modified most of the interfaces and their return objects and some entity classes, and added websocket。修改了大部分接口以及其返回对象和部分实体类，并新增了websocket

+++ src/main/java/com/share/controller/CommentController.java
@@ -1,13 +1,16 @@
 package com.share.controller;
 
+import com.share.entity.Comments;
 import com.share.exceptions.ShareException;
 import com.share.result.RestObject;
 import com.share.result.RestResponse;
-import com.share.ro.CommentsRo;
+import com.share.ro.commentsRo.ArticleCommentsRo;
+import com.share.ro.commentsRo.CommentsRo;
+import com.share.ro.commentsRo.TravelCommentsRo;
+import com.share.ro.commentsRo.UserCommentsRo;
 import com.share.service.CommentsService;
 import com.share.vo.CommentsVo;
 import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiModelProperty;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -34,16 +37,35 @@ public class CommentController {
         }
     }
 
-    @ApiOperation("删除评论")
-    @PostMapping("/deleteComments/{id}")
-    public RestObject<String> deleteComments(@PathVariable int id){
-        commentService.deleteComments(id);
-        return RestResponse.makeOKRsp("删除成功!");
+    @ApiOperation("删除评论；id为评论id；userId为当前登录用户id")
+    @PostMapping("/deleteComments/{id}/{userId}")
+    public RestObject<String> deleteComments(@PathVariable int id,@PathVariable int userId){
+        Comments comments = commentService.queryCommentsById(id);
+        Integer userIdCommentator = comments.getUserIdCommentator();
+        if (userIdCommentator == userId){
+            commentService.deleteComments(id);
+            return RestResponse.makeOKRsp("删除成功!");
+        }else {
+            return RestResponse.makeErrRsp("删除失败！您不是该条评论的主人哦");
+        }
     }
 
-    @ApiOperation("按用户id、出行id、信息id查询评论")
-    @GetMapping("/queryById")
-    public RestObject<List<CommentsRo>> queryById(@RequestBody CommentsVo commentsVo){
-        return RestResponse.makeOKRsp(commentService.queryCommentsById(commentsVo));
+    @ApiOperation("按用户id查询评论")
+    @GetMapping("/queryByUserId/{userId}")
+    public RestObject<List<UserCommentsRo>> queryByUserId(@PathVariable int userId){
+        return RestResponse.makeOKRsp(commentService.queryCommentsByUserId(userId));
     }
+
+    @ApiOperation("按出行id查询评论")
+    @GetMapping("/queryByTravelId/{travelId}")
+    public RestObject<List<TravelCommentsRo>> queryByTravelId(@PathVariable int travelId){
+        return RestResponse.makeOKRsp(commentService.queryCommentsByTravelId(travelId));
+    }
+
+    @ApiOperation("按信息id查询评论")
+    @GetMapping("/queryByArticleId/{articleId}")
+    public RestObject<List<ArticleCommentsRo>> queryByArticleId(@PathVariable int articleId){
+        return RestResponse.makeOKRsp(commentService.queryCommentsByArticleId(articleId));
+    }
+
 }
